Move some exceptions into .errors
